requirements.txt
psycopg2-binary
llama-index==0.9.48
torch
transformers
python-dotenv
asyncpg
sqlalchemy[asyncio]
sqlalchemy
pgvector
fastapi[standard]
